fix: move parsing logic to FromStr Bootnode trait
