+++ ethrex/src/main.rs
@@ -1,9 +1,9 @@
 use core::types::Genesis;
 use net::types::BootNode;
-use std::num::ParseIntError;
 use std::{
     io::BufReader,
     net::{AddrParseError, SocketAddr},
+    str::FromStr,
 };
 use tokio::join;
 use tracing::Level;
@@ -49,7 +49,8 @@ async fn main() {
     let bootnode_string = matches
         .get_one::<String>("bootnodes")
         .expect("bootnode is required");
-    let bootnode = parse_bootnode(&bootnode_string);
+
+    let bootnode = BootNode::from_str(bootnode_string).expect("Failed to parse bootnode");
 
     let http_socket_addr =
         parse_socket_addr(http_addr, http_port).expect("Failed to parse http address and port");
@@ -77,23 +78,4 @@ fn read_genesis_file(genesis_file_path: &str) -> Genesis {
 
 fn parse_socket_addr(addr: &str, port: &str) -> Result<SocketAddr, AddrParseError> {
     format!("{addr}:{port}").parse()
-}
-
-fn parse_bootnode(input: &str) -> BootNode {
-    // TODO: error handling
-    let node_id = decode_hex(&input[8..136]).unwrap();
-    let socket_address: SocketAddr = input[137..]
-        .parse()
-        .expect("Failed to parse bootnode address and port");
-    BootNode {
-        node_id,
-        socket_address,
-    }
-}
-
-pub fn decode_hex(s: &str) -> Result<Vec<u8>, ParseIntError> {
-    (0..s.len())
-        .step_by(2)
-        .map(|i| u8::from_str_radix(&s[i..i + 2], 16))
-        .collect()
 }
